core: huge mailbox refactor
changed over all the currently used mailbox messages to structs of their
own. i'm not super happy with th API right now, but at least the fields
are named.

also had to refactor the tests to for ensuring the structs have the
right layout, and...i'm *slightly* concerned about how trustworthy they
are.

but that's something i'll have to test on real hardware probably.