core: add virtual/physical/bus address types
to make `MemoryMapper`...worthwhile, added in `ArmAddress<A>`, where `A`
is either `Physical` or `Virtual`, and `BusAddress` which wraps a `u32`
representing a...bus address.

keeping track of that by hand is extremely error prone, and while
implementing this, i realized i'd completely screwed this up in the code
for sending mailbox messages for initialization.

i'm still not entirely convinced that's right, but that needs to be
tested on real hardware.