core: cleanup + `Debug` impl
removed a panic if the framebuffer size is 0. technically, a user could
want a buffer of size 0, however stupid that may be.

some sort of error checking to ensure that what's returned is actually
what was asked for is probably a good idea, but at least we're not
panicking for now.

+ `Debug` on `Context` and related types to help with checking things
are working right.